+++ JudoPayDotNet/Http/Connection.cs
@@ -14,8 +14,6 @@ using Newtonsoft.Json.Serialization;
 
 namespace JudoPayDotNet.Http
 {
-    using JudoPayDotNet.Models;
-
     /// <summary>
     /// Handles the http requests creation and the http responses
     /// </summary>
@@ -88,12 +86,6 @@ namespace JudoPayDotNet.Http
                 }
             }
 
-            var paymentModel = body as PaymentModel;
-            if (paymentModel != null && !string.IsNullOrWhiteSpace(paymentModel.UserAgent))
-            {
-                request.Headers.UserAgent.TryParseAdd(paymentModel.UserAgent);
-            }
-            
             if (body != null)
             {
                 request.Content = new StringContent(JsonConvert.SerializeObject(body, _settings), new UTF8Encoding());